added pipeline handler for infer + fixed faces index selection in assign_poster
